ARROW-856: Also read compiler info from stdout
As I cannot reproduce the problem, this may not the be the correct solution. Still due to the output of `gcc -v` in the ticket, it is very likely that some systems print the info on stdout. Also the additional message should improve debugging.

Author: Uwe L. Korn <uwelk@xhochy.com>

Closes #650 from xhochy/ARROW-856 and squashes the following commits:

d87c7bd [Uwe L. Korn] ARROW-856: Also read compiler info from stdout